backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
werkzeug==3.1.3
python-dotenv==1.1.0
PyJWT==2.10.1
PyMuPDF==1.25.3
requests==2.32.3
